+++ common/src/main/java/com/yanny/ali/datagen/LanguageHolder.java
@@ -23,57 +23,74 @@ public class LanguageHolder {
         TRANSLATION_MAP.put("ali.description.count_bonus", "%s (%s %s)");
         TRANSLATION_MAP.put("ali.description.quality", "Quality: %s");
 
-        TRANSLATION_MAP.put("ali.type.condition.all_of", "All must pass:");
-        TRANSLATION_MAP.put("ali.type.condition.and", "And:");
-        TRANSLATION_MAP.put("ali.type.condition.any_biome_check", "Any Biome Check:");
-        TRANSLATION_MAP.put("ali.type.condition.any_dimension", "Any Dimension:");
+        /*
+         * PREDICATES
+         */
+
+        TRANSLATION_MAP.put("ali.type.condition.all_of", "All Of:");
         TRANSLATION_MAP.put("ali.type.condition.any_of", "Any of:");
-        TRANSLATION_MAP.put("ali.type.condition.biome_check", "Biome Check:");
         TRANSLATION_MAP.put("ali.type.condition.block_state_property", "Block State Property:");
-        TRANSLATION_MAP.put("ali.type.condition.can_tool_perform_action", "Can Tool Perform Action: %s");
-        TRANSLATION_MAP.put("ali.type.condition.contains_loot", "Contains Loot:");
-        TRANSLATION_MAP.put("ali.type.condition.custom_param_predicate", "Custom Param Predicate:");
-        TRANSLATION_MAP.put("ali.type.condition.damage_source", "Damage Source:");
         TRANSLATION_MAP.put("ali.type.condition.damage_source_properties", "Damage Source Properties:");
         TRANSLATION_MAP.put("ali.type.condition.entity_properties", "Entity Properties:");
         TRANSLATION_MAP.put("ali.type.condition.entity_scores", "Entity Scores:");
         TRANSLATION_MAP.put("ali.type.condition.inverted", "Inverted:");
-        TRANSLATION_MAP.put("ali.type.condition.is_light_level", "Is Light Level:");
-        TRANSLATION_MAP.put("ali.type.condition.killed_by_player", "Must be killed by player");
+        TRANSLATION_MAP.put("ali.type.condition.killed_by_player", "Killed by player");
         TRANSLATION_MAP.put("ali.type.condition.location_check", "Location Check:");
-        TRANSLATION_MAP.put("ali.type.condition.loot_table_id", "Loot Table Id: %s");
-        TRANSLATION_MAP.put("ali.type.condition.main_hand_table_bonus", "Main Hand Table Bonus:");
-        TRANSLATION_MAP.put("ali.type.condition.match_equipment_slot", "Match Equipment Slot:");
-        TRANSLATION_MAP.put("ali.type.condition.match_fluid", "Match Fluid:");
-        TRANSLATION_MAP.put("ali.type.condition.match_killer_distance", "Match Killer Distance:");
-        TRANSLATION_MAP.put("ali.type.condition.match_player", "Match Player:");
         TRANSLATION_MAP.put("ali.type.condition.match_tool", "Match Tool:");
-        TRANSLATION_MAP.put("ali.type.condition.not", "Not:");
-        TRANSLATION_MAP.put("ali.type.condition.or", "Or:");
-        TRANSLATION_MAP.put("ali.type.condition.player_param_predicate", "Player Param Predicate:");
         TRANSLATION_MAP.put("ali.type.condition.random_chance", "Random Chance:");
         TRANSLATION_MAP.put("ali.type.condition.random_chance_with_looting", "Random Chance With Looting:");
         TRANSLATION_MAP.put("ali.type.condition.reference", "Reference: %s");
-        TRANSLATION_MAP.put("ali.type.condition.survives_explosion", "Must survive explosion");
+        TRANSLATION_MAP.put("ali.type.condition.survives_explosion", "Survives Explosion");
         TRANSLATION_MAP.put("ali.type.condition.table_bonus", "Table Bonus:");
         TRANSLATION_MAP.put("ali.type.condition.time_check", "Time Check:");
         TRANSLATION_MAP.put("ali.type.condition.value_check", "Value Check:");
         TRANSLATION_MAP.put("ali.type.condition.weather_check", "Weather Check:");
 
+        // Forge
+        TRANSLATION_MAP.put("ali.type.condition.can_tool_perform_action", "Can Tool Perform Action: %s");
+        TRANSLATION_MAP.put("ali.type.condition.loot_table_id", "Loot Table Id: %s");
+
+        //TODO match_entity -> podla lootjs zdrojaku su to 3 rozne podmienky!!!!!!
+
+        // LootJS
+        TRANSLATION_MAP.put("ali.type.condition.and", "And:");
+        TRANSLATION_MAP.put("ali.type.condition.any_biome", "Any Biome:");
+        TRANSLATION_MAP.put("ali.type.condition.any_dimension", "Any Dimension:");
+        TRANSLATION_MAP.put("ali.type.condition.any_structure", "Any Structure:");
+        TRANSLATION_MAP.put("ali.type.condition.biome", "Biome:");
+        TRANSLATION_MAP.put("ali.type.condition.block_predicate", "Block Predicate:");
+        TRANSLATION_MAP.put("ali.type.condition.direct_killer_predicate", "Direct Killer Predicate:");
+        TRANSLATION_MAP.put("ali.type.condition.distance_to_killer", "Distance To Killer:");
+        TRANSLATION_MAP.put("ali.type.condition.entity_predicate", "Entity Predicate:");
+        TRANSLATION_MAP.put("ali.type.condition.killer_predicate", "Killer Predicate:");
+        TRANSLATION_MAP.put("ali.type.condition.light_level", "Light Level:");
+        TRANSLATION_MAP.put("ali.type.condition.match_damage_source", "Match Damage Source:");
+        TRANSLATION_MAP.put("ali.type.condition.match_equipment_slot", "Match Equipment Slot:");
+        TRANSLATION_MAP.put("ali.type.condition.match_loot", "Match Loot:");
+        TRANSLATION_MAP.put("ali.type.condition.match_mainhand", "Match Mainhand:");
+        TRANSLATION_MAP.put("ali.type.condition.match_offhand", "Match Offhand:");
+        TRANSLATION_MAP.put("ali.type.condition.match_player", "Match Player:");
+        TRANSLATION_MAP.put("ali.type.condition.not", "Not:");
+        TRANSLATION_MAP.put("ali.type.condition.or", "Or:");
+        TRANSLATION_MAP.put("ali.type.condition.player_predicate", "Player Predicate:");
+        TRANSLATION_MAP.put("ali.type.condition.random_chance_with_enchantment", "Random Chance With Enchantment:");
+
+        /*
+         * MODIFIERS
+         */
+
         TRANSLATION_MAP.put("ali.type.function.apply_bonus", "Apply Bonus:");
         TRANSLATION_MAP.put("ali.type.function.copy_name", "Copy Name:");
         TRANSLATION_MAP.put("ali.type.function.copy_nbt", "Copy Nbt:");
         TRANSLATION_MAP.put("ali.type.function.copy_state", "Copy State:");
-        TRANSLATION_MAP.put("ali.type.function.custom_player", "Custom Player Modifier:");
         TRANSLATION_MAP.put("ali.type.function.enchant_randomly", "Enchant Randomly:");
         TRANSLATION_MAP.put("ali.type.function.enchant_with_levels", "Enchant With Levels:");
         TRANSLATION_MAP.put("ali.type.function.exploration_map", "Exploration Map:");
         TRANSLATION_MAP.put("ali.type.function.explosion_decay", "Explosion Decay");
         TRANSLATION_MAP.put("ali.type.function.fill_player_head", "Fill Player Head:");
-        TRANSLATION_MAP.put("ali.type.function.furnace_smelt", "Use Smelting Recipe On Item");
+        TRANSLATION_MAP.put("ali.type.function.furnace_smelt", "Furnace Smelt");
         TRANSLATION_MAP.put("ali.type.function.limit_count", "Limit Count:");
         TRANSLATION_MAP.put("ali.type.function.looting_enchant", "Looting Enchant:");
-        TRANSLATION_MAP.put("ali.type.function.modified_item", "Modified dynamically! UNKNOWN RESULT!");
         TRANSLATION_MAP.put("ali.type.function.reference", "Reference:");
         TRANSLATION_MAP.put("ali.type.function.set_attributes", "Set Attributes:");
         TRANSLATION_MAP.put("ali.type.function.set_banner_pattern", "Set Banner Pattern:");
@@ -88,6 +105,10 @@ public class LanguageHolder {
         TRANSLATION_MAP.put("ali.type.function.set_nbt", "Set Nbt:");
         TRANSLATION_MAP.put("ali.type.function.set_potion", "Set Potion:");
         TRANSLATION_MAP.put("ali.type.function.set_stew_effect", "Set Stew Effect:");
+
+        // LootJS
+        TRANSLATION_MAP.put("ali.type.function.custom_player", "Custom Player Modifier:");
+        TRANSLATION_MAP.put("ali.type.function.modified_item", "Modified dynamically!");
 
         TRANSLATION_MAP.put("ali.enum.group_type.all", "Selects all entries");
         TRANSLATION_MAP.put("ali.enum.group_type.random", "Selects random entry");