+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed sporadic missing loot info
 - Fixed sporadic crash on startup
 - Fixed show recipe outputs doesn't follow modifier changes
+- Changed modifiers and predicate names to match official names
 
 ## [1.2.0]